% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{dr_frost_account}
\alias{dr_frost_account}
\title{Store your account details securely}
\usage{
dr_frost_account(email, keyring = NULL)
}
\arguments{
\item{email}{The email for your Dr Frost Account. Multiple accounts with
different emails can be stored.}

\item{keyring}{A string, passed into \code{\link[keyring:key_get]{keyring::key_set()}} and
\code{\link[keyring:key_get]{keyring::key_get()}}. Use this to store your Dr Frost credentials on a
custom keyring.}
}
\value{
Nothing. This function is called for its side effects.
}
\description{
Securely store the email and password to your Dr Frost Maths account so that
it can be used between sessions. The
\link[keyring]{keyring::\code{keyring-package}} package is used as a backend,
meaning that credentials are stored securely and permanently.
}
\details{
Calling this function should open a prompt, allowing you to enter your
password securely.

To change the password stored for a specific email, just call the function
again with the same email.

If the account you store is the first to be stored on the specified keyring,
the account will be marked as a 'default' for the keyring. This means that
when \code{\link[=perform_tasks]{perform_tasks()}} is called without specifying an email, this account
will be used as a default. Change this with \code{\link[=set_default_account]{set_default_account()}}.

This means that once you have stored an account, the \code{\link[=perform_tasks]{perform_tasks()}}
function can be called without specifying an email or password.
}
\examples{
\dontrun{
  dr_frost_account("YOUR-EMAIL")
  perform_tasks()
}

}
