% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_frost_tasks.R
\name{dr_frost_tasks}
\alias{dr_frost_tasks}
\title{Get the set of available tasks}
\usage{
dr_frost_tasks(include = NULL, exclude = NULL)
}
\arguments{
\item{include}{A character vector of tasks to include. If \code{NULL}, all tasks
will be included.}

\item{exclude}{A character vector of tasks to exclude. If \code{NULL}, no tasks
will be excluded.}

\item{x}{A \code{dr_frost_tasks} object.}

\item{...}{Not used.}
}
\value{
A data frame with additional class 'dr_frost_tasks', where each row
describes a task. The table has 4 variables:
\describe{
\item{topic}{The topic that the task falls under.}
\item{subtopic}{The subtopic of the task.}
\item{task_name}{The name of the task.}
\item{description}{A brief description of the task.}
}
}
\description{
Retrieve a set of tasks that can be completed in
\href{https://www.drfrostmaths.com/}{Dr Frost Maths}, choosing tasks to include
and exclude.
}
\details{
If the \code{include} argument matches a topic, subtopic or task name of any task,
tasks will require a direct match with the string to be included. Otherwise,
partial matching/pattern recognition will be used to identify matches.

The same applies to the \code{exclude} argument.

While this makes specifying tasks simple most of the time, it can sometimes
lead to some unexpected cases where more complex searches result in less
matches.
}
\examples{
dr_frost_tasks()

dr_frost_tasks(include = "addition_subtraction")

dr_frost_tasks(include = "Arithmetic Operations")

dr_frost_tasks(include = "KS3 Number")

dr_frost_tasks(include = "KS3", exclude = "Arithmetic")

dr_frost_tasks(exclude = "addition")

tibble::as_tibble(dr_frost_tasks())

}
\seealso{
\code{\link[=perform_tasks]{perform_tasks()}}
}
