% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{install_selene}
\alias{install_selene}
\title{Install the selene python package}
\usage{
install_selene(method = "auto", conda = "auto")
}
\arguments{
\item{method, conda}{Passed into \code{\link[reticulate:py_install]{reticulate::py_install()}}}
}
\description{
tidydrfrost depends on the 'selene' python package. This function resolves
this dependency.
}
\examples{
\dontrun{
  install_selene()
}

}
