% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{install_tdf_dependencies}
\alias{install_tdf_dependencies}
\title{Install Python dependencies for tidydrfrost}
\usage{
install_tdf_dependencies(method = "auto", conda = "auto")
}
\arguments{
\item{method, conda}{Passed into \code{\link[reticulate:py_install]{reticulate::py_install()}}}
}
\description{
tidydrfrost depends on a few python packages. This function installs these
using \link[reticulate:py_install]{reticulate::py_install}.
}
\details{
tidydrfrost currently depends on 'selene', 'sigfig', 'sympy' and 'latex2sympy2'.
For added control over the installation of these packages, use
\code{\link[reticulate:py_install]{reticulate::py_install()}} directly:

\if{html}{\out{<div class="sourceCode">}}\preformatted{reticulate::py_install(c("selene", "sigfig", "sympy", "latex2sympy2"))
}\if{html}{\out{</div>}}

Or use \code{pip}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{pip install selene sigfig sympy latex2sympy2
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
  install_tdf_dependencies()
}

}
