% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_tasks.R
\name{perform_tasks}
\alias{perform_tasks}
\title{Complete Dr Frost Maths tasks}
\usage{
perform_tasks(
  tasks = dr_frost_tasks(),
  email = NULL,
  password = NULL,
  keyring = NULL,
  end = TRUE
)
}
\arguments{
\item{tasks}{The tasks to complete. Either a \code{\link[=dr_frost_tasks]{dr_frost_tasks()}} object, or
a character vector. If a character vector is supplied, it will be passed
as the \code{include} argument into the \code{\link[=dr_frost_tasks]{dr_frost_tasks()}} function. By default,
all available tasks are completed.}

\item{email}{Used to login to the website. If \code{NULL}, the default email is
used.}

\item{password}{Used to login to the website. If \code{NULL}, the function will
try to find the password corresponding to the \code{email} that has been set
using \code{\link[=dr_frost_account]{dr_frost_account()}}.}

\item{keyring}{A string, passed into \code{\link[keyring:key_get]{keyring::key_get()}}. Use this if your
Dr Frost credentials are stored on a keyring that is not the default.}

\item{end}{Whether the browser session should be closed when the tasks are
completed.}
}
\value{
A data frame of information about each task, with 4 variables:
\describe{
\item{task}{The task that was attempted.}
\item{completed}{Whether the task was successfully completed.}
\item{points}{The number of points that were earned by completing the task.}
\item{error}{If the task failed, the error object that was returned.}
}
}
\description{
Complete a set of \href{https://www.drfrostmaths.com/}{Dr Frost Maths} tasks. Each
task rewards the user with a certain number of points, which can vary.
}
\details{
Completing Dr Frost Tasks requires a continuous (although not necessarily
fast) internet connection, and tasks will fail if no connection is found for
an extended period of time.

Completing each task calls a corresponding function in Python.
}
\examples{
\dontrun{
  perform_tasks("addition")
  perform_tasks()
}

}
\seealso{
\itemize{
\item \code{\link[=dr_frost_account]{dr_frost_account()}} to store your account details securely.
\item \code{\link[=install_selene]{install_selene()}} to install python dependencies.
}
}
