% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_frost_tasks.R
\name{print.dr_frost_tasks}
\alias{print.dr_frost_tasks}
\title{Printing Dr Frost tasks}
\usage{
\method{print}{dr_frost_tasks}(x, ..., n = 10)
}
\arguments{
\item{x}{A \code{dr_frost_tasks} object.}

\item{...}{Not used.}

\item{n}{The maximum number of tasks to show. Excess tasks will be truncated.
Use \code{Inf} to show all tasks.}
}
\value{
\code{x}, invisibly
}
\description{
The \code{dr_frost_tasks} object has an enhanced printing method, using topics and
subtopics as headers.
}
\details{
This printing method is heavily inspired by the tibble printing method:
\link[tibble:formatting]{tibble::print.tbl_df}.
}
\examples{
print(dr_frost_tasks())
print(dr_frost_tasks(), n = 5)

}
