% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_frost_account.R
\name{set_default_account}
\alias{set_default_account}
\title{Set the default Dr Frost Maths account}
\usage{
set_default_account(email, keyring = NULL)
}
\arguments{
\item{email}{The email to use.}

\item{keyring}{A string, passed into \code{\link[keyring:key_get]{keyring::key_set()}} and
\code{\link[keyring:key_get]{keyring::key_get()}}. Use this if your Dr Frost accounts are
stored on a custom keyring.}
}
\value{
Nothing. This function is called for its side effects.
}
\description{
Set or change the default email used by the \code{\link[=perform_tasks]{perform_tasks()}} function.
}
\examples{
\dontrun{
  set_default_account("YOUR-EMAIL")
}

}
