% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydrfrost-package.R
\docType{package}
\name{tidydrfrost-package}
\alias{tidydrfrost}
\alias{tidydrfrost-package}
\title{tidydrfrost: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ashbythorpe/tidydrfrost}
  \item Report bugs at \url{https://github.com/ashbythorpe/tidydrfrost/issues}
}

}
\author{
\strong{Maintainer}: Ashby Thorpe \email{ashbythorpe@gmail.com} (\href{https://orcid.org/0000-0003-3106-099X}{ORCID})

}
\keyword{internal}
