% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydrfrost-package.R
\docType{package}
\name{tidydrfrost-package}
\alias{tidydrfrost}
\alias{tidydrfrost-package}
\title{tidydrfrost: Automated Tasks and Point Farmer for Dr Frost Maths}
\description{
Automates a set of tasks in Dr Frost Maths (https://www.drfrostmaths.com/). Provides a robust and secure login method and a detailed Command Line Interface to make the completing tasks easy and robust.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ashbythorpe/tidydrfrost}
  \item \url{https://ashbythorpe.github.io/tidydrfrost/}
  \item Report bugs at \url{https://github.com/ashbythorpe/tidydrfrost/issues}
}

}
\author{
\strong{Maintainer}: Ashby Thorpe \email{ashbythorpe@gmail.com} (\href{https://orcid.org/0000-0003-3106-099X}{ORCID})

}
\keyword{internal}
